+++ Assets/Scripts/Services/Foo.cs
@@ -2,6 +2,6 @@ namespace MrWatts.Internal.FuelInject.TestProject
 {
     public sealed class Foo
     {
-        private readonly Bar bar;
+        private Bar Bar { get; set; }
     }
 }