+++ Assets/Package/CHANGELOG.md
@@ -1,3 +1,12 @@
-# 0.1.0 (Not Released Yet)
+# x.y.z (Not Released Yet)
+
+-   Internal restructuring.
+-   `IInjector<object>`, `IInjector<GameObject>`, and `IInjector<Scene>` are now available for clients to use when they need to inject dependencies into objects at runtime (e.g. after the scene was already loaded).
+
+# 0.1.1
+
+-   Fix some meta file GUIDs conflicting with Onion.
+
+# 0.1.0
 
 -   Initial release.